+++ src/main/java/lotto/model/Money.java
@@ -7,35 +7,35 @@ import lotto.exception.InvalidMoneyException;
 
 public class Money {
 
-    private final int money;
+    private final long money;
 
-    public Money(final int money) {
+    public Money(final long money) {
         validateOverThousand(money);
         validateDivideByThousand(money);
         this.money = money;
     }
 
-    private void validateOverThousand(final int money) {
+    private void validateOverThousand(final long money) {
         if (!isOverThousand(money)) {
             throw new InvalidMoneyException(money);
         }
     }
 
-    private boolean isOverThousand(final int money) {
+    private boolean isOverThousand(final long money) {
         return money >= LOTTO_PRICE.getValue();
     }
 
-    private void validateDivideByThousand(final int money) {
+    private void validateDivideByThousand(final long money) {
         if (!canDivideByThousand(money)) {
             throw new InvalidMoneyException(money);
         }
     }
 
-    private boolean canDivideByThousand(final int money) {
+    private boolean canDivideByThousand(final long money) {
         return money % LOTTO_PRICE.getValue() == ZERO.getValue();
     }
 
-    public int getMoney() {
+    public long getMoney() {
         return money;
     }
 }
